+++ backend/hankkimoa/src/main/java/com/jjangiji/hankkimoa/restaurant/controller/RestaurantController.java
@@ -1,13 +1,15 @@
 package com.jjangiji.hankkimoa.restaurant.controller;
 
 import com.jjangiji.hankkimoa.auth.config.AuthRequiredPrincipal;
+import com.jjangiji.hankkimoa.restaurant.service.RecommendService;
 import com.jjangiji.hankkimoa.restaurant.service.RestaurantBatchService;
 import com.jjangiji.hankkimoa.restaurant.service.RestaurantService;
 import com.jjangiji.hankkimoa.restaurant.service.dto.reqeust.RecommendationFeedbackRequest;
-import com.jjangiji.hankkimoa.restaurant.service.dto.response.RestaurantSimpleResponse;
 import com.jjangiji.hankkimoa.restaurant.service.dto.reqeust.RestaurantCreateRequest;
+import com.jjangiji.hankkimoa.restaurant.service.dto.response.RecommendServerRestaurantsResponse;
 import com.jjangiji.hankkimoa.restaurant.service.dto.response.RestaurantResponse;
 import com.jjangiji.hankkimoa.restaurant.service.dto.response.RestaurantSearchResponse;
+import com.jjangiji.hankkimoa.restaurant.service.dto.response.RestaurantSimpleResponse;
 import com.jjangiji.hankkimoa.user.domain.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -23,6 +25,7 @@ import java.util.List;
 @RestController
 public class RestaurantController {
 
+    private final RecommendService recommendService;
     private final RestaurantBatchService restaurantBatchService;
     private final RestaurantService restaurantService;
 
@@ -35,7 +38,7 @@ public class RestaurantController {
     @PostMapping("/api/recommendation/feedback")
     public ResponseEntity<Void> createRecommendationFeedback(@AuthRequiredPrincipal User user,
                                                              @RequestBody RecommendationFeedbackRequest request) {
-        restaurantService.createRecommendationFeedback(user, request);
+        recommendService.createRecommendationFeedback(user, request);
         return ResponseEntity.noContent().build();
     }
 
@@ -47,7 +50,8 @@ public class RestaurantController {
 
     @GetMapping("/api/recommendation/restaurants")
     public ResponseEntity<List<RestaurantSimpleResponse>> readRecommendRestaurants(@AuthRequiredPrincipal User user) {
-        List<RestaurantSimpleResponse> recommendRestaurants = restaurantService.readRecommendRestaurants(user);
+        RecommendServerRestaurantsResponse response = recommendService.readRecommendRestaurants(user);
+        List<RestaurantSimpleResponse> recommendRestaurants = restaurantService.readRestaurants(user, response.uniqueIds());
         return ResponseEntity.ok(recommendRestaurants);
     }
