+++ backend/hankkimoa/src/main/java/com/jjangiji/hankkimoa/restaurant/service/dto/response/RecommendServerRestaurantsResponse.java
@@ -2,5 +2,5 @@ package com.jjangiji.hankkimoa.restaurant.service.dto.response;
 
 import java.util.List;
 
-public record RecommendServerRestaurantsResponse(List<String> restaurantUniqueIds) {
+public record RecommendServerRestaurantsResponse(List<String> uniqueIds) {
 }
